xc refactor : 1. put all tests in one file 2. remove some unnecessary public: in header
